refactor: complete project re-structuring
