+++ src/drawthis/gui/shaders/basic.frag
@@ -1,0 +1,10 @@
+#version 330
+
+in vec2 v_uv;
+out vec4 frag_color;
+
+uniform sampler2D tex;
+
+void main() {
+    frag_color = texture(tex, v_uv);
+}
