--- src/drawthis/render/__init__.py
@@ -1,9 +1,0 @@
-from .feh_backend import start_slideshow_feh
-from .opengl_backend import start_slideshow_ogl
-from .slideshow_manager import SlideshowManager
-
-__all__ = [
-    "start_slideshow_feh",
-    "start_slideshow_ogl",
-    "SlideshowManager",
-]
