+++ src/tests/unit/persistence/test_logic_database_sqlite3.py
@@ -1,0 +1,232 @@
+import unittest
+from sqlite3 import ProgrammingError
+
+from drawthis.core.models.session.dataclasses import ImageRow
+from drawthis.persistence.session.sqlite3_backend import SQLite3Backend
+
+"""
+Behavioral contract for SQLite3Backend:
+
+Core behaviors:
+  - Inserts rows in provided batch via insert_rows() in single transaction
+  - Removes rows for provided parent paths in single transaction
+  - Maintains schema integrity and migrations*
+    *Additive migrations only
+
+Data guarantees:
+  - Deterministic-random ordering of rows across reads* by randid
+  - Seen status persists correctly on close
+  - Writes and reads are atomic and committed
+  - Provides metrics of inserted/removed rows
+    *Writes overwrite the randid changing order
+
+Error handling:
+  - Handles recoverable errors (locked DB, invalid inputs) gracefully
+"""
+
+
+def nested_list():
+    batch = [
+        [
+            ImageRow(file_path=f"{parent}/{i}.png", randid=i, mtime=i)
+            for i in range(5)
+        ]
+        for parent in {"ching", "bing", "deng_xiao_ping"}
+    ]
+    return batch
+
+
+def flat_list():
+    batch = [
+        ImageRow(file_path=f"{parent}/{i}.png", randid=i, mtime=i)
+        for i in range(5)
+        for parent in ["ching", "bing", "deng_xiao_ping"]
+    ]
+    return batch
+
+
+class TestInsertBehavior(unittest.TestCase):
+    """
+    Behavioral contract: Insert ops must always...
+
+    Assumptions:
+      - Every row is a valid ImageRow instance
+      - Provided data is a flat iterable (including single rows)
+
+    Compromises:
+      - Uniqueness defined as ≥2 of {path, inode+dev, size} matching
+
+    Expected behavior:
+    EB1 Files with identical path & size → skipped
+    EB2 Files with identical inode+dev & size → skipped
+    EB3 Files with identical path & inode+dev → skipped
+    EB4 Files with fewer than 2 matches → inserted as unique
+
+    Acceptance criteria:
+    AC1 insert_rows() returns count of unique inserts
+    AC2 Database contains no duplicate paths after insertion
+    AC3 Insertion commits atomically (WAL checkpoint visible)
+    """
+
+    def test_input_all_flat_iterable_types(self):
+        cases = [
+            ("list", lambda flat: flat, 1, 15),
+            ("tuple", tuple, 1, 15),
+            ("set", set, 1, 15),
+            ("iterator", iter, 1, 15),
+            ("generator", lambda flat: (row for row in flat), 1, 15),
+        ]
+        for iter_t, conv, exp_commits, exp_rows in cases:
+            with self.subTest(
+                iter_type=iter_t,
+                converter=conv,
+                exp_rows=exp_rows,
+                exp_commits=exp_commits,
+            ):
+                # select fixture
+                fixture = nested_list if "nested" in iter_t else flat_list
+
+                # convert to desired iterable
+                batch_to_test = conv(fixture())
+
+                # create backend and insert
+                backend = SQLite3Backend(db_path=":memory:")
+                print(f"Testing: {iter_t}")
+
+                rows = backend.insert_rows(batch_to_test)
+                self.assertEqual(exp_rows, rows)
+
+    def test_input_all_nested_iterable_types(self):
+        """
+        Demonstrates that nested iterables are *not supported*.
+        Backend expects flat iterables only. Nested lists or generators must
+        be flattened or inserted one nested object at a time, as shown.
+        """
+        cases = [
+            ("nested_list", lambda nested: nested, 15),
+            (
+                "nested_generator",
+                lambda nested: (iter(row) for row in nested),
+                15,
+            ),
+        ]
+
+        for iter_type, converter, exp_rows in cases:
+            with self.subTest(iter_type=iter_type):
+                # select fixture
+                fixture_func = (
+                    nested_list if "nested" in iter_type else flat_list
+                )
+                nested_iterable = converter(fixture_func())
+
+                # create backend
+                backend = SQLite3Backend(db_path=":memory:")
+                print(f"Testing: {iter_type}")
+
+                # directly inserting nested objects should raise
+                with self.assertRaises(ProgrammingError):
+                    backend.insert_rows(nested_iterable)
+
+                # regenerate iterable for actual insertion
+                fresh_iterable = converter(fixture_func())
+
+                # insert each nested object individually
+                rows_inserted = 0
+                for row_group in fresh_iterable:
+                    rows_inserted += backend.insert_rows(row_group)
+
+                self.assertEqual(exp_rows, rows_inserted)
+
+    def test_same_path_and_size(self):
+        self.skipTest("Spec not implemented yet")
+
+    def test_same_inode_plus_dev_and_size(self):
+        self.skipTest("Spec not implemented yet")
+
+    def test_same_name_and_inode_plus_dev(self):
+        self.skipTest("Spec not implemented yet")
+
+    def test_unique_insert_count_return(self):
+        self.skipTest("Spec not implemented yet")
+
+    def test_commit_and_checkpoint_after_insert(self):
+        self.skipTest("Spec not implemented yet")
+
+
+class TestRemoveBehavior(unittest.TestCase):
+    """
+    Behavioral contract: removal ops must always...
+
+    Assumptions:
+      - Removal criteria is a field in a row, thus contained in an ImageRow
+      - Provided data is iterable or singular
+
+    Expected behavior:
+    EB1 Every entry that matches removal criteria should be removed
+        (regardless of other fields)
+    EB2 All entries that have not been removed are unaltered
+    EB3 Row IDs are reordered to maintain exact order of remaining rows
+
+    Acceptance criteria:
+    AC1 Database contains no matching entries after removal
+    AC2 Remaining entries are unaltered (except for primary ID)
+    AC3 Partial matches are not altered
+    AC4 Each removal query (by provided criteria) is atomic
+    """
+
+    def test_matching_entries_removed(self):
+        """EB1 Remove every entry that matches"""
+        self.skipTest("Spec not implemented yet")
+
+    def test_remaining_rows_are_unaltered(self):
+        """EB2 All entries that remain are unaltered"""
+        self.skipTest("Spec not implemented yet")
+
+    def test_row_id_preserves_old_order_without_gaps(self):
+        """EB3 Row IDs are reordered"""
+        self.skipTest("Spec not implemented yet")
+
+    def test_no_partial_match_removed(self):
+        """AC3 Partial matches are not altered"""
+        self.skipTest("Spec not implemented yet")
+
+    def test_commit_and_checkpoint_after_removal(self):
+        """AC4 Each removal query is atomic"""
+        self.skipTest("Spec not implemented yet")
+
+
+class TestLoadBehavior(unittest.TestCase):
+    """
+    Behavioral contract: Load operations must always...
+
+    Assumptions:
+      - Data consistency is guaranteed within a single load operation
+        (no mid-read mutations)
+      - Row order is preserved for any accepted criterion
+
+    Expected behavior:
+    EB1 All entries in the database are returned
+    EB2 Entries are returned sorted by the provided criterion
+    EB3 Entries can be deterministically indexed by that criterion, and
+        the order must match the load order
+
+    Acceptance criteria:
+    AC1 All entries are provided, sorted according to a deterministic criterion
+    AC2 Ordering and indexing are reproducible given the same criterion
+    """
+
+    def test_every_entry_returned(self):
+        """EB1 All entries in the database are returned"""
+        self.skipTest("Spec not implemented yet")
+
+    def test_entries_sorted(self):
+        """EB2 Entries are returned sorted by the provided criterion"""
+        self.skipTest("Spec not implemented yet")
+
+    def test_deterministic_sort_order(self):
+        """EB3 Entries can be deterministically indexed"""
+        self.skipTest("Spec not implemented yet")
+
+    def test_index_reproducible_by_criterion(self):
+        """AC2 Ordering and indexing are reproducible"""
+        self.skipTest("Spec not implemented yet")
